[gokillctl] add trigger view
